requirements.txt
streamlit
pandas
sqlalchemy
pyodbc
python-dotenv
streamlit-aggrid
openpyxl
xlsxwriter
reportlab